bots/requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
python-dotenv==0.20.0
emoji==2.0.0
